% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basesurv.R
\name{basesurv}
\alias{basesurv}
\title{Compute the estimations of values of the survival related functions}
\usage{
basesurv(time, status, preds, times.eval = NULL, centered = FALSE)
}
\arguments{
\item{time}{the prediction variables}

\item{status}{individual observation times}

\item{preds}{the predictions of the log-hazard function in CoxLGBM model}

\item{times.eval}{evaluation time points, The default is \code{NULL}}

\item{centered}{the default is \code{FALSE}}
}
\value{
a \code{list} of time points, and the values of cumulative baseline
hazard function and baseline survival function at these time points
}
\description{
Creates a \code{list} of values of cumulative baseline hazard
function and baseline survival function at evaluation time points.
}
