% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_nloglik.R
\name{cox.nloglik}
\alias{cox.nloglik}
\title{Calculate the value of loss function}
\usage{
cox.nloglik(preds, dtrain)
}
\arguments{
\item{preds}{the predictions of the log-hazard function in CoxLGBM model}

\item{dtrain}{a \code{\link[lightgbm]{lgb.Dataset}} object}
}
\value{
the value of loss function
}
\description{
Calculate the value of loss function
}
