% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxlgbm_cv.R
\name{coxlgbm.cv}
\alias{coxlgbm.cv}
\title{Main CV logic for CoxLGBM}
\usage{
coxlgbm.cv(time, status, x, obj = lossobj, eval = cox.nloglik, ...)
}
\arguments{
\item{time}{individual observation times}

\item{status}{censoring indicators}

\item{x}{the prediction variables}

\item{obj}{objective function, It should be set to \code{lossobj} here}

\item{eval}{evaluation metric, It should be set to \code{cox.nloglik} here}

\item{...}{other parameters for \code{lightgbm}}
}
\value{
a trained model \code{lgb.CVBooster}
}
\description{
Main CV logic for CoxLGBM
}
