% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_imp.R
\name{coxlgbm.importance}
\alias{coxlgbm.importance}
\title{Compute feature importance in a model}
\usage{
coxlgbm.importance(model, percentage = TRUE)
}
\arguments{
\item{model}{object of class \code{lgb.Booster}}

\item{percentage}{whether to show importance in relative percentage}
}
\value{
a \code{data.table} of feature importances in a model
}
\description{
Compute feature importance in a model
}
