% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{coxlgbm.predict}
\alias{coxlgbm.predict}
\title{Model Predictions}
\usage{
coxlgbm.predict(
  object,
  time,
  status,
  newx,
  times.eval = NULL,
  centered = FALSE
)
}
\arguments{
\item{object}{a trained CoxLGBM model}

\item{time}{the new individual observation times}

\item{status}{the new censoring indicators}

\item{newx}{the new prediction variables}

\item{times.eval}{evaluation time points, The default is \code{NULL}}

\item{centered}{the default is \code{FALSE}}
}
\value{
a \code{list} of time points, the values of cumulative baseline
hazard function and baseline survival function at these time points,
and the predicted log-hazard function
}
\description{
Model Predictions
}
