% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{coxlgbm.train}
\alias{coxlgbm.train}
\title{Train a CoxLGBM model}
\usage{
coxlgbm.train(time, status, x, obj = lossobj, eval = cox.nloglik, ...)
}
\arguments{
\item{time}{individual observation times}

\item{status}{censoring indicators}

\item{x}{the prediction variables}

\item{obj}{objective function, It should be set to \code{lossobj} here}

\item{eval}{evaluation metric, It should be set to \code{cox.nloglik} here}

\item{...}{other parameters for \code{lightgbm}}
}
\value{
a trained \code{lgb.Booster}
}
\description{
Train a CoxLGBM model
}
\examples{
# Generate some survival data with 10 informative covariates
n <- 200; p <- 20
beta <- c(rep(1,10),rep(0,p-10))
x <- matrix(rnorm(n*p),n,p)
real.time <- -(log(runif(n)))/(10*exp(drop(x \%*\% beta)))
cens.time <- rexp(n,rate=1/10)
status <- ifelse(real.time <= cens.time,1,0)
obs.time <- ifelse(real.time <= cens.time,real.time,cens.time)
# Fit a CoxLGBM model
model <- coxlgbm.train(time = obs.time, status = status, x = x, obj = lossobj, eval = cox.nloglik,
                       params = list(min_data = 5, learning_rate = 0.05), eval_freq = 0, nrounds = 100)
}
