% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_obj.R
\name{lossobj}
\alias{lossobj}
\title{Calculate the first and second order gradients}
\usage{
lossobj(preds, dtrain)
}
\arguments{
\item{preds}{the predictions of the log-hazard function in CoxLGBM model}

\item{dtrain}{a \code{\link[lightgbm]{lgb.Dataset}} object}
}
\value{
a list of the first and second order gradients
}
\description{
Calculate the first and second order gradients
}
